gwt: Unify UploadWidget(s).
