+++ gwt/vmchecker-gui/src/ro/pub/cs/vmchecker/client/model/Assignment.java
@@ -1,15 +1,22 @@
 package ro.pub.cs.vmchecker.client.model;
 
 public class Assignment {
+	public enum StorageType {
+		NORMAL,
+		LARGE
+	};
+
 	public String id;
 	public String title;
-	public String storageType;
+	public StorageType storageType;
 	public String storageHost;
 	public String storageBasepath;
 	public String deadline;
 	public String statementLink;
 
-	public Assignment(String id, String title, String storageType, String storageHost, String storageBasepath, String deadline, String statementLink) {
+	public Assignment(String id, String title, StorageType storageType,
+			String storageHost, String storageBasepath,
+			String deadline, String statementLink) {
 		this.id = id;
 		this.title = title;
 		this.storageType = storageType;
